Add FConstraintBarrier and FLockBarrier, and related functionality
Note that the hierarchical responsibilities in the Constraint Barrier Pattern currently differs slightly from the Shape Barrier Pattern; The constraint barrier base class is the only one to hold the NativeRef, and the subclasses are only responsible of creating and assigning the correct NativeRef->Native, and then just type-cast it whenever temporarily needed. The design got much simpler by doing so, in my opinion.