Distance Constraint disables Lock Controller in Details Panel, and never writes to the native Lock Controller
This is necessary because the implementation of agx::DistanceJoint seem to use the Lock Controller internally itself (because it is enabled implicitly after having created the constraint), so if the Lock Controller is written to by us the standard behaviour of the constraint becomes broken.

Note that there might be other special constraints that have similar requirements, so we should make a generic system for hiding of specific controllers per constraint type (or the reverse: enabling of controllers)!