Improve FAGX_SimpleMeshSceneProxy vertex buffer population
FAGX_SimpleMeshSceneProxy now populates the vertex buffers itself instead of using deprecated helper function InitFromDynamicVertex(). This is a necessary preparation for adding support for explicit normals, texture coordinates, and index buffer. One intermediate buffer is also eliminated with this change, making it faster, and it is a step towards making further optimizations possible.