Fix so that Constraint Frames visualization does not block interaction with the transform gizmo
The way this is fixed is simply by making sure that the lines used to visualize constraint frames are always thinner than those of the transform gizmo, so that the transform gizmo is thickest in the scenario where they are overlapping. This scenario should be solvable by other means (HitProxy, depth sorting, etc), but it did not seem to work in the particular scenario when the transform gizmo and coordinate frame visualization are within a mesh while the viewport camera is outside the mesh (i.e. occlusion by the mesh).