Don't call MarkRenderDynamicDataDirty from CreateSceneProxy if currently ticking
That's not allowed and leads to a crash. Haven't found any official
documentation about this though, so the details are very unclear. I did
find an example where another plugin does something similar and calls
SendRenderDynamicData_Concurrent in this case, so we do the same for
now.

More study needed.