Enable OpenVDB linking crash workaround
It's been back-and-forth on this one serveral times now. Sometimes a
particular configuration works, and sometimes it doesn't. Today I must
link with at least OpenVDB to avoid crashes when creating an AGX
Terrain. I'm committing this to avoid future crashes, but I hope we'll
be able to find a proper long-term fix for the crashes somehow. This
approach is at least OK since it contains no hard coded search paths,
explicit dlopens or unexplainable function calls. To have
openvdb::initialize in AGXUnreal is a bit... odd... but not entirely
repulsive.