Do not create a CollisionGroupManager from DisabledCollisionGroupsComponent
Instead try again later if one doesn't exist, until the first
TickCompoinent callback.

On-demand spawning of a new CollisionGroupManager sometimes caused
multiple CollisionGroupManagers to be created. Not entirely sure why.

It is now the user's responsibility to create a CollisionGroupManager. I
don't like it much, so this may be investigated further in the future.