Add PostInitProperties and SkipUCSModifiedProperties to TwoBodyTire
PostInitProperties because we now set Owning Actor there instead of in
PostLoad.
SkipUCSModifiedProperties because we want the Owning Actor to be
modifiable on instances of Blueprints containing tires, which they
aren't by default because Unreal Editor considers property changes
during setup (constructor, PostInitProperties, PostLoad, ...) to be part
of the Construction Script and thus protected properties.