Fix AGX Dynamics body and shape objects being duplicated and invalidated on Details Panel changes
The source of the problem is that editing a Component in a Blueprint
instance through the Details Panel causes all Components in that
Blueprint to be recreated. The recreation happens through a serialize,
destroy, create, deserialize sequence. This change uses Component
Instance Data to store the address of the AGX Dynamics objects in the
serialization and GIsReconstructingBlueprintInstance to avoid destroying
and later creating AGX Dynamics objects at such times. Conceptually this
causes the AGX Dynamics object to be moved from the old Component to the
new via the serialization storage.

To avoid a lot of code duplication, one for each type of AGX Dynamics
object we have, IAGX_NativeOwner is introduced. This interface provides
the basic operations needed to transfer ownership of an AGX Dynamics
object from one Component instance to another. Each Component type that
wraps an AGX Dynamics object should implement IAGX_NativeOwner and
implement the Uneal Engine function GetComponentInstanceData. See Rigid
Body Component and Shape Component for examples.