Fix Wire Node type ComboBox bug when switching node
The ComboBox remembers it's selection from the previous node, making it
impossible to switch to that node type since the ComboBox believes it's
already selected. Fixed by comparing the current selection with the
expected in UpdateValues and switching if needed. Must do the check
because doing the switch unconditionally breaks the ComboBox.