In the contact materials test, remove trailing 's' from imported asset folders to delete
Because the naming convention was changed with the URDF import branch
since StaticMeshs, StaticMesh + s, isn't correct. The contact material
test was never updated and this wasn't discovered because this test
hasn't been run for a while because of a typo in the test group name,
which has now also been fixed.