Call the Property Changed callback on all selected objects
Not just the one that got the PostEditChangeChainProperty member
function called. This is required because a bug in Unreal Engine causes
some of the objects to not have their member function called when
multiple Components in a Blueprint instance is selected and edited.

A side effect of this is that it is no longer allowed to have
object-specific state captured in the callback, any object's callback
can now be used to update any other object of the same type.