Remove LoadPreviouslyImportedRenderMaterials, add Render Material registering in UpdateAndSaveRenderMaterialAsset
There is no need to explicitly load and register all Render Material
assets up front because the loop over them in AddOrUpdateRenderMaterials
will visit precisely the assets that are still on drive since we already
deleted those that were on drive but not among the simulation objects.
Regardless of whether the asset is created or updated in that loop, it
will be registered in the RestoredRenderMaterials collection.