Remove constexpr from InvalidAngleType
Because constexpr expressions have stricter correctness requirements
than regular code and does not allow an enum variable to take a value
other than those listed in the enum definition. We want to force a -1
value to signal an error. We should find a better way to do this, proper
error reporting.