Add option to pass soil particle data by-ID to Niagara
By-ID meaning that particles are stored at indices decided by their
entity ID instead of by their entity index. This means that there may be
gaps in the arrays. Get Terrain Particle Data Array has been updated to
handle this case, but not yet with an easy-to-flick switch to toggle
between by-index arrays and by-ID arrays. Visibility tag is used to tell
the mesh renderer which particles should be rendered or not.