Unregister Shovel Component from Shovel Properties in End Play instead of destructor
Fixes freeze in the garbage collector after ending second Play In Editor
session. End Play is better than the destructor for cleanup since it is
run deterministically at the end of a session, while the destructor can
run at any later point in time. Especially important when dealing with
garbage collecting smart pointers, which we do here, since we don't want
garbage collection to mess with smart pointers. Not sure if that last
bit is true, I'm trying to find an explanation for the freeze.