Remove old FAGX_RigidBodyReference, rename FAGX_BodyReference to FAGX_RigidBodyReference
A swap, basically. The old Rigid Body Reference has been replaced with
the new one everywhere. Backwards compatiblity is handled with a core
redicted for the change in name from BodyName to Name in AGXUnreal.cpp.
It is either sheer luck that we can replace the type like this, or it
may be that it is actually broken and another strategy must be found.
Testing will tell.

In the process we lost support for caching. I never trusted it, so we're
better of without it until we can find a way to have safe caching. Don't
remember exactly what the problem with caching was, why a
TWeakObjectPointer to invalidate it wasn't enough. It may be related to
the Constraint tab in the AGX Mode panel, i.e. where constraints can be
created to bodies in the level. It is not OK to have pointers from the
UI into the level, apparently. The error message is

    Illegal TEXT reference to a private object in external package