Reset global dynamic wire contacts state after unit test
Exposed bEnableWireWireCollisions and bEnableDynamicWireContacts to the
plugin settings, so we have a default to reset to. Also marked a bunch
of other UAGX_Simulation Properties BlueprintReadOnly, no reason why
they should be inaccessable. Some might even be BlurprintReadWrite, but
making sure changes made at runtime are updated on the AGX Dynamics
state is out of scope for this branch.