Use FAGX_NotificationUtilities::ShowNotification to show no-licence message also on Linux
Before we had a message box that didn't behave correctly on Linux so
that platform fell back to other means of displaying the error, means
that aren't included in shipping builds. Then we changed Windows to use
a notification instead of a message box but kept the separate code path
for Linux. This change removes the separate code path so all platforms
use ShowNotification.