Add SkipUCSModifiedProperties to Wire Component > Route Nodes
Needed because Route Nodes now contains a Component Reference that is
initialized in Post Init Properties. That is considered part of the
Construction Script which means that the property and all its outer
properties including Route Nodes are marked as authorative, i.e. should
not be serialized during Blueprint Reconstruction. This is wrong in our
case since we want the user to be able to edit Route Nodes and its
members and we do want those changes to survive Blueprint
Reconstruction. This desire is communicated to Unreal Engine by adding
the Skip UCS Modified Properties meta flag to the property, Route Nodes
on this case.