No extra overload of Set Motion Control for Blueprint
Cannot have a separate UAGX_RigidBody::SetMotionControl_BP with a
default value because that is not backward compatible with Blueprints
that already use the Set Motion Control node. Setting the Display Name
equal is not enough to make the Blueprint Compiler find the function.

Instead using the Meta Function Specififer to set the default value.