Revert "Merge branch 'feature/ball_joint_twist_range_controller' into 'master'"
This reverts commit 7093ee4540acb6dbe8ea1092b7133fcd5320fd36, reversing
changes made to 2a6e8b927d68c562b30343a52cb6e07c73d341a0.

The Twist Range Controller branch did too much at once and when it was
discovered that it causes build errors on Windows for some Unreal Engine
versions it turned out to be difficult to identify what caused the error.  The
entire merge is reverted and the constituent change sets are created and merged
as separate branches instead.