When creating Scene Components during model import and synchronization use the owning Actor as the outer instead of the attach parent
Because Unreal Engine does inital unique name enforcement on the outer
level first and then on the Actor level at a later time. This means that
for a while during import and synchronization we may have name
conflicts, and later we may have tables that contain names that aren't
correct anymore because a Component was renamed due to a name collision.